requirements.txt
aiogram==3.4.1
httpx==0.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9